tty bug with multi-line prompts
This effectively reverts ee95f0e00bb071e658bf605fe2bdc8ba01ecc145
because it was interfering with prompts that spanned more than one
line (by inserting '--More-- + erase' between displaying of prompt and
getting input for the answer).

So we're back to the situation where autodescribe feedback when moving
the cursor will leave text on the second line if it generates text too
wide for one line.  (^R redraws the screen correctly.)