Guidebook.mn Symbols section
The table listing symset entries was ok for ps/pdf output but for
plain text output it was wrapping to a second line for S_xan,
S_pet_override, and S_player_override and also going past the right
margin for some others.  I've shortened the text for those three
and narrowed the first column to give more room for the third column,
bringing all the lines within their margin.

The entry for S_stone was rather surprising:  "dark part of a room".
It now says "solid rock or unexplored terrain or dark part of a room"
and needed to span two lines.

The first column has been labeled "Default" but that was why that
column was wider than necessary.  It now has a label of two spaces
and is left justified instead of being centered.  It is also now
forcing constant-width Roman as the font and every character in that
column should be the same size so one reason for centering is removed.

The title used to be centered but always looked a bit off due to the
fact that the three columns have differing widths and/or because of
the over-wide lines changing the location of "center".  It is now
left justified and looks a little bit like it's labeling the first
column.  Perhaps it needs a blank line separating it from the rest.

Guidebook.tex is lagging again.